Quick Start/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.9
sqlparse==0.5.3